[VIEW-MODEL] Refactor
