+++ app/src/main/java/com/example/prueba001/composable/detail/DetailComposable.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.prueba001.R
-import com.example.prueba001.bbdd.viewmodel.DataBaseViewModel
+import com.example.prueba001.viewModels.DataBaseViewModel
 import com.example.prueba001.composable.customViews.TopBackBar
 import com.example.prueba001.model.HeroModel
 import com.example.prueba001.utils.mapToDb
