+++ app/src/main/java/com/example/prueba001/composable/navigation/NavigationController.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.prueba001.bbdd.viewmodel.DataBaseViewModel
+import com.example.prueba001.viewModels.DataBaseViewModel
 import com.example.prueba001.composable.detail.DetailComposable
 import com.example.prueba001.composable.list.ListComposable
 import com.example.prueba001.model.HeroModel
